+++ x-pack/plugins/threat_intelligence/public/modules/indicators/components/field_value/field_value.tsx
@@ -10,7 +10,7 @@ import { useFieldTypes } from '../../hooks/use_field_types';
 import { EMPTY_VALUE } from '../../../../common/constants';
 import { Indicator, RawIndicatorFieldId } from '../../../../types/indicator';
 import { DateFormatter } from '../../../../components/date_formatter';
-import { unwrapValue } from '../../lib/unwrap_value';
+import { unwrapValue } from '../../utils/unwrap_value';
 
 export interface IndicatorFieldValueProps {
   /**
