+++ x-pack/plugins/threat_intelligence/public/modules/indicators/components/flyout/flyout.tsx
@@ -24,7 +24,7 @@ import { FormattedMessage } from '@kbn/i18n-react';
 import { InvestigateInTimelineButton } from '../../../timeline/components/investigate_in_timeline_button';
 import { DateFormatter } from '../../../../components/date_formatter/date_formatter';
 import { Indicator, RawIndicatorFieldId } from '../../../../types/indicator';
-import { unwrapValue } from '../../lib/unwrap_value';
+import { unwrapValue } from '../../utils/unwrap_value';
 import { IndicatorFlyoutJsonTab } from './components/json_tab';
 import { IndicatorFlyoutTableTab } from './components/table_tab';
 import { IndicatorFlyoutOverviewTab } from './components/overview_tab';
