+++ x-pack/plugins/threat_intelligence/public/modules/indicators/components/table/components/cell_actions.tsx
@@ -11,7 +11,7 @@ import { ComponentType } from '../../../../../types/component_type';
 import { Indicator } from '../../../../../types/indicator';
 import { Pagination } from '../../../hooks/use_indicators';
 import { AddToTimeline } from '../../../../timeline/components/add_to_timeline';
-import { fieldAndValueValid, getIndicatorFieldAndValue } from '../../../lib/field_value';
+import { fieldAndValueValid, getIndicatorFieldAndValue } from '../../../utils/field_value';
 import { FilterIn } from '../../../../query_bar/components/filter_in';
 import { FilterOut } from '../../../../query_bar/components/filter_out';
