+++ x-pack/plugins/threat_intelligence/public/modules/indicators/utils/get_field_schema.ts
