+++ x-pack/plugins/threat_intelligence/public/modules/query_bar/components/filter_out/filter_out.tsx
@@ -10,8 +10,11 @@ import { EuiButtonEmpty, EuiButtonIcon, EuiContextMenuItem, EuiToolTip } from '@
 import { Filter } from '@kbn/es-query';
 import { ComponentType } from '../../../../types/component_type';
 import { useIndicatorFiltersContext } from '../../../indicators/hooks/use_filters_context';
-import { fieldAndValueValid, getIndicatorFieldAndValue } from '../../../indicators/lib/field_value';
-import { FilterOut as FilterOutConst, updateFiltersArray } from '../../lib/filter';
+import {
+  fieldAndValueValid,
+  getIndicatorFieldAndValue,
+} from '../../../indicators/utils/field_value';
+import { FilterOut as FilterOutConst, updateFiltersArray } from '../../utils/filter';
 import { Indicator } from '../../../../types/indicator';
 import { useStyles } from './styles';
 import { ICON_TITLE } from './translations';
