+++ x-pack/plugins/threat_intelligence/public/modules/timeline/hooks/use_investigate_in_timeline.ts
@@ -8,10 +8,10 @@
 import { useContext } from 'react';
 import moment from 'moment';
 import { DataProvider } from '@kbn/timelines-plugin/common';
-import { generateDataProvider } from '../lib/data_provider';
+import { generateDataProvider } from '../utils/data_provider';
 import { SecuritySolutionContext } from '../../../contexts/security_solution_context';
-import { fieldAndValueValid, getIndicatorFieldAndValue } from '../../indicators/lib/field_value';
-import { unwrapValue } from '../../indicators/lib/unwrap_value';
+import { fieldAndValueValid, getIndicatorFieldAndValue } from '../../indicators/utils/field_value';
+import { unwrapValue } from '../../indicators/utils/unwrap_value';
 import {
   Indicator,
   IndicatorFieldEventEnrichmentMap,
